refactor: выделить интерфейсы UserValidator и FilmValidator для улучшения абстракции
- Создать интерфейсы UserValidator и FilmValidator для определения контракта валидации
- Существующиt UserValidator и FilmValidator переименовать в UserValidatorImpl и FilmValidatorImpl для ясности
- Сервисный слой теперь зависит от интерфейса, а не реализации
- Сохранена обратная совместимость с существующей логикой валидации
